chore: move v2 plugin into old folder
