+++ inlang/source-code/plugins/inlang-message-format/src/index.ts
@@ -1,2 +1,2 @@
-import { plugin } from "./plugin.js"
+import { plugin } from "./v2/plugin.js"
 export default plugin
